requirements.txt
asgiref==3.9.1
asttokens==3.0.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
decorator==5.2.1
Django==5.2.5
django-filter==25.1
django-phonenumber-field==8.1.0
django-unfold==0.65.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
executing==2.2.0
idna==3.10
iniconfig==2.1.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Markdown==3.8.2
matplotlib-inline==0.1.7
packaging==25.0
parso==0.8.4
pexpect==4.9.0
phonenumberslite==9.0.12
pillow==11.3.0
pluggy==1.6.0
prompt_toolkit==3.0.51
psycopg==3.2.10
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.3.4
python-decouple==3.8
requests==2.32.4
requests-mock==1.12.1
setuptools==75.8.0
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.15.0
urllib3==2.5.0
wcwidth==0.2.13
zarinpal-py-sdk==0.1.4
